Add check if the button is held for the whole 5 seconds
If the button is held down to reboot the watch and not released after the reboot, the device can enter recovery mode unintentionally. This check aims to ignore the button if it is held for the entirety of the 5 second delay, implying it was held before the device started to boot and should be ignored.